Fix post links
