+++ tests/pages/components/post-view.js
@@ -21,7 +21,8 @@ function imageUrl(elementSelector) {
     css("background-image").
     replace(/['"]+/g, '');
 
-  const cloudinaryUrl = backgroundImageStyle.substring(4, backgroundImageStyle.length - 1);
+  const cloudinaryUrl = backgroundImageStyle.
+    substring(4, backgroundImageStyle.length - 1);
 
-  return cloudinaryUrl.replace("https://res.cloudinary.com/cloudinary-test/image/fetch/w_1920/", "");
+  return cloudinaryUrl.substring(cloudinaryUrl.lastIndexOf("http"));
 }
